Updated code to remove warnings

+++ src/Application.cpp
@@ -13,28 +13,28 @@ int main(int argc, char *argv[])
 	po::store(po::parse_command_line(argc, argv, desc), vm);
 	po::notify(vm);
 
-	// // float circleRadius = 60.0;
-	std::vector points { Point<float>({ 10, 100 }),
-		                 Point<float>({ 10.2f, 300 }),
-		                 Point<float>({ 10.6f, 500 }),
-		                 Point<float>({ 11, 700 }),
-		                 Point<float>({ 11.4f, 900 }),
-		                 Point<float>({ 12.1f, 1000 }),
-		                 Point<float>({ 12.3f, 1200 }),
-		                 Point<float>({ 12.7f, 1400 }),
-		                 Point<float>({ 13.3f, 1600 }),
-		                 Point<float>({ 13.7f, 1710 }),
-		                 Point<float>({ 14, 1700 }),
-		                 Point<float>({ 14.3f, 1600 }),
-		                 Point<float>({ 14.7f, 1500 }),
-		                 Point<float>({ 15.1f, 1400 }),
-		                 Point<float>({ 15.7f, 1300 }) };
+	// // double circleRadius = 60.0;
+	std::vector points { Point<double>({ 10, 100 }),
+		                 Point<double>({ 10.2, 300 }),
+		                 Point<double>({ 10.6, 500 }),
+		                 Point<double>({ 11, 700 }),
+		                 Point<double>({ 11.4, 900 }),
+		                 Point<double>({ 12.1, 1000 }),
+		                 Point<double>({ 12.3, 1200 }),
+		                 Point<double>({ 12.7, 1400 }),
+		                 Point<double>({ 13.3, 1600 }),
+		                 Point<double>({ 13.7, 1710 }),
+		                 Point<double>({ 14, 1700 }),
+		                 Point<double>({ 14.3, 1600 }),
+		                 Point<double>({ 14.7, 1500 }),
+		                 Point<double>({ 15.1, 1400 }),
+		                 Point<double>({ 15.7, 1300 }) };
 	// std::vector<Point> points { Point({ 10, 10 }),
 	// 	                        Point({ 20, 20 }),
 	// 	                        Point({ 30, 30 }),
 	// 	                        Point({ 40, 40 }),
 	// 	                        Point({ 50, 50 }) };
-	std::vector<Point<float>> points2(360);
+	std::vector<Point<double>> points2(360);
 	for (int i = 0; i < points2.size();i++) {
 		points2[i].x = i;
 		points2[i].y = sin(i * 3.14 / 180);
